feat: improve converter hierarchy and split it into scalar converters and conditional converters
